Add some typemap to the Python swig interface.
This these new typemap, you can now do some basic stuff with histograms.
I will commit a demo for this soon...

The _cv.cpp and _highgui.cpp generated files are also updated. The new version
of the _highgui.cpp also correct a problem with the new signature of the
function icvInitAVIWriter()


git-svn-id: https://code.ros.org/svn/opencv/trunk@169 73c94f0f-984f-4a5f-82bc-2d8db8d8ee08