Add support of V4L2_PIX_FMT_MJPEG.
With this patch, new webcams of Logitech, like QuickCam Fusion works.
Note: For use these webcams, look at the UVC driver at
http://linux-uvc.berlios.de/

Decompression of the MJPEG stream is done via a new local class MyMJpegReader,
derivated from GrFmtJpegReader. The only change is to use fmemopen() instead of
fopen(). This mean, this part of code is only usable with __USE_GNU.


git-svn-id: https://code.ros.org/svn/opencv/trunk@444 73c94f0f-984f-4a5f-82bc-2d8db8d8ee08