Another patch from "Phil" on the OpenCV Mailing list:
In tsysa.cpp, the relative path for the testdata directory should be two
subdirectory levels about the module location.  This is because the cvtest in
the src directory is a libtool wrapper.  The real binary is in the .libs
directory below the src directory.

Note: this don't works "as is" when you do the compilation in a separate
directory.


git-svn-id: https://code.ros.org/svn/opencv/trunk@450 73c94f0f-984f-4a5f-82bc-2d8db8d8ee08