Patch from Kertesz Csaba:
#1: It solves a value range problem by such a v4l2 image properties which have
too narrow value range: e.g. 0-7

#2: General bug in my last patch. v4l (VIDIOCGPICT) gives back the default
image properties and changes the webcam to the default state.  There is no
feedback from v4l if user modifies the image properties of a device. The
functionality after these modification is normal: OpenCV returns with the
default or last adjusted value of an image property.


git-svn-id: https://code.ros.org/svn/opencv/trunk@451 73c94f0f-984f-4a5f-82bc-2d8db8d8ee08