Move the test of --with-gtk and --with-carbon later.
Configure will test with_gtk=yes only after we have decided if we have carbon
or not. This will prevent a Linux user without carbon, and specifing --with-gtk
to have problems.
The main behavior on Mac OS X is not changed: without --with-carbon and
--with-gtk, it use carbon, same with only --with-carbon. With --with-gtk, it
cause an error. To have gtk on Mac OS X, you need to say --with-gtk and
--without-carbon.
Close sourceforge bug #1554473.


git-svn-id: https://code.ros.org/svn/opencv/trunk@870 73c94f0f-984f-4a5f-82bc-2d8db8d8ee08